+++ GradimoZajedno.Web/gradimo-zajedno/src/scss/layout/_forms.scss
@@ -14,6 +14,14 @@ textarea {
 	&:focus {
 		outline: 1px dashed $purple-dark;
 	}
+	&:invalid {
+		border: none;
+		background-color: transparent;
+	}
+	&:focus:invalid {
+		outline: none;
+		border: 2px solid $red;
+	}
 }
 
 input[type='text'],
@@ -39,7 +47,8 @@ input[type='password'] {
 	line-height: 19px;
 	letter-spacing: 0.04em;
 }
-.input {
+input.input,
+textarea.input {
 	border: 2px solid $purple;
 	border-radius: 8px;
 	width: 100%;
@@ -49,7 +58,4 @@ input[type='password'] {
 	line-height: 19px;
 	letter-spacing: 0.04em;
 	padding: 6px;
-}
-.input:invalid {
-	border: 2px solid $red;
 }
