+++ GradimoZajedno.Web/wwwroot/index.html
@@ -460,12 +460,12 @@
 			<div class="forms__content">
 				<div class="forms__col">
 					<h2 class="h1 forms__title">Želite da stupite u kontakt sa nama? Pošaljite nam poruku!</h2>
-					<form class="form js-form" action="/">
+					<form class="form js-form" action="/contact">
 						<div class="form__input-wrap">
 							<div class="input-wrap">
 								<label for="ime" class="label">Ime:</label>
 								<span class="error-msg"></span>
-								<input class="input js-input" type="text" name="ime" id="ime">
+								<input class="input js-input" type="text" name="ime" required id="ime" autocomplete="off">
 								<span class="form__error-msg js-error-msg"></span>
 							</div>
 						</div>
@@ -473,7 +473,7 @@
 							<div class="input-wrap">
 								<label for="prezime" class="label">Prezime:</label>
 								<span class="error-msg"></span>
-								<input class="input js-input" type="text" name="prezime" id="prezime">
+								<input class="input js-input" type="text" name="prezime" required id="prezime" autocomplete="off">
 								<span class="form__error-msg js-error-msg"></span>
 							</div>
 						</div>
@@ -481,7 +481,7 @@
 							<div class="input-wrap">
 								<label for="email" class="label">Email:</label>
 								<span class="error-msg"></span>
-								<input class="input js-input" type="email" name="email" id="email">
+								<input class="input js-input" type="email" name="email" required id="email" autocomplete="off">
 								<span class="form__error-msg js-error-msg"></span>
 							</div>
 						</div>
@@ -489,13 +489,13 @@
 							<div class="input-wrap">
 								<label for="tel" class="label">Kontakt telefon:</label>
 								<span class="error-msg"></span>
-								<input class="input js-input" type="tel" name="tel" id="tel">
+								<input class="input js-input" type="tel" name="tel" id="tel" autocomplete="off">
 								<span class="form__error-msg js-error-msg"></span>
 							</div>
 						</div>
 						<div class="form__input-wrap form__input-wrap--full">
 							<label for="poruka" class="label">Vaša poruka:</label>
-							<textarea class="input js-input" name="poruka" id="poruka" cols="30" rows="5"></textarea>
+							<textarea class="input js-input" name="poruka" id="poruka" cols="30" rows="5" required></textarea>
 							<span class="form__error-msg"></span>
 						</div>
 						<div class="form__btn-wrap">
