+++ GradimoZajedno.Web/wwwroot/single-object.html
@@ -245,12 +245,12 @@
 							<div class="wrap">
 								<div class="forms__content">
 									<div class="forms__col">
-										<form class="form js-form" action="/">
+										<form class="form js-form" action="/contact">
 											<div class="form__input-wrap">
 												<div class="input-wrap">
 													<label for="ime" class="label">Ime:</label>
 													<span class="error-msg"></span>
-													<input class="input js-input" type="text" name="ime" id="ime">
+													<input class="input js-input" type="text" name="ime" id="ime" autocomplete="off">
 													<span class="form__error-msg js-error-msg"></span>
 												</div>
 											</div>
@@ -258,7 +258,7 @@
 												<div class="input-wrap">
 													<label for="prezime" class="label">Prezime:</label>
 													<span class="error-msg"></span>
-													<input class="input js-input" type="text" name="prezime" id="prezime">
+													<input class="input js-input" type="text" name="prezime" id="prezime" autocomplete="off">
 													<span class="form__error-msg js-error-msg"></span>
 												</div>
 											</div>
@@ -266,7 +266,7 @@
 												<div class="input-wrap">
 													<label for="email" class="label">Email:</label>
 													<span class="error-msg"></span>
-													<input class="input js-input" type="email" name="email" id="email">
+													<input class="input js-input" type="email" name="email" id="email" autocomplete="off">
 													<span class="form__error-msg js-error-msg"></span>
 												</div>
 											</div>
@@ -274,7 +274,7 @@
 												<div class="input-wrap">
 													<label for="tel" class="label">Kontakt telefon:</label>
 													<span class="error-msg"></span>
-													<input class="input js-input" type="tel" name="tel" id="tel">
+													<input class="input js-input" type="tel" name="tel" id="tel" autocomplete="off">
 													<span class="form__error-msg js-error-msg"></span>
 												</div>
 											</div>
